+++ src/boostConfiguration.ts
@@ -131,6 +131,16 @@ export class BoostConfiguration {
             Defaults.analysisModelValue;
     }
 
+    public static get refreshAnalysisAlways(): boolean {
+        return (workspace.getConfiguration(NOTEBOOK_TYPE, null).get(Defaults.refreshAnalysisAlwaysName) as boolean)??
+            Defaults.refreshAnalysisAlwaysValue;
+    }
+
+    public static refreshAnalysisAlwaysByKernel(kernel: string): boolean {
+        return (workspace.getConfiguration(NOTEBOOK_TYPE, null).get(kernel + "." + Defaults.refreshAnalysisAlwaysName) as boolean)??
+            Defaults.refreshAnalysisAlwaysValue;
+    }
+
     static _cachedVersion: string = "";
     public static get version(): string | undefined {
         if (this._cachedVersion) {
@@ -197,4 +207,7 @@ class Defaults {
 
     public static readonly analysisModelName : string = "analysisModel";
     public static readonly analysisModelValue : string = "";
+
+    public static readonly refreshAnalysisAlwaysName : string = "refreshAnalysisAlways";
+    public static readonly refreshAnalysisAlwaysValue : boolean = false;
 }